backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.31",
]
